Changed floats to doubles in sect.  Theoretically, we may have got imprecise mean coverage estimates when using really big sequences, such as whole chromosome length.

+++ src/sect/sect.hpp
@@ -73,8 +73,8 @@ namespace kat
         StringSet<CharString>           names;
         StringSet<Dna5String>           seqs;
         vector<vector<uint64_t>*>       *counts;    // K-mer counts for each K-mer window in sequence (in same order as seqs and names; built by this class)
-        vector<float>                   *coverages; // Overall coverage calculated for each sequence from the K-mer windows.
-        vector<float>                   *gcs;       // GC% for each sequence
+        vector<double>                  *coverages; // Overall coverage calculated for each sequence from the K-mer windows.
+        vector<double>                  *gcs;       // GC% for each sequence
         vector<uint32_t>                *lengths;   // Length in nucleotides for each sequence
 
         int                             resultCode;
@@ -284,8 +284,8 @@ namespace kat
         void createBatchVars(uint16_t batchSize)
         {
             counts = new vector<vector<uint64_t>*>(batchSize);
-            coverages = new vector<float>(batchSize);
-            gcs = new vector<float>(batchSize);
+            coverages = new vector<double>(batchSize);
+            gcs = new vector<double>(batchSize);
             lengths = new vector<uint32_t>(batchSize);
         }
 
@@ -428,7 +428,7 @@ namespace kat
 
             uint64_t seqLength = seq.length();
             uint64_t nbCounts = seqLength - kmer + 1;
-            float mean_cvg = 0;
+            double mean_cvg = 0.0;
 
             if (seqLength < kmer)
             {
@@ -464,7 +464,7 @@ namespace kat
                 (*counts)[index] = seqCounts;
 
                 // Assumes simple mean calculation for sequence coverage for now... plug in Bernardo's method later.
-                mean_cvg = (float)sum / (float)nbCounts;
+                mean_cvg = (double)sum / (double)nbCounts;
                 (*coverages)[index] = mean_cvg;
 
             }
@@ -489,21 +489,19 @@ namespace kat
                     ns++;
             }
 
-            float gc_perc = ((float)(gs + cs)) / ((float)(seqLength - ns));
+            double gc_perc = ((double)(gs + cs)) / ((double)(seqLength - ns));
             (*gcs)[index] = gc_perc;
 
-            float log_cvg = args->cvg_logscale ? log10(mean_cvg) : mean_cvg;
+            double log_cvg = args->cvg_logscale ? log10(mean_cvg) : mean_cvg;
 
             // Assume log_cvg 5 is max value
-            float compressed_cvg = args->cvg_logscale ? log_cvg * (args->cvg_bins / 5.0) : mean_cvg * 0.1;
+            double compressed_cvg = args->cvg_logscale ? log_cvg * (args->cvg_bins / 5.0) : mean_cvg * 0.1;
 
-            uint16_t x = gc_perc * args->gc_bins;  // Convert float to 1.dp
+            uint16_t x = gc_perc * args->gc_bins;  // Convert double to 1.dp
             uint16_t y = compressed_cvg >= args->cvg_bins ? args->cvg_bins - 1 : compressed_cvg;      // Simply cap the y value
 
             // Add bases to matrix
             contamination_mx->getThreadMatrix(th_id)->inc(x, y, seqLength);
-
         }
-
     };
 }
